Move useColors to index file
