+++ frontend/src/components/ColumnChartPreview.tsx
@@ -9,7 +9,7 @@ import {
   CartesianGrid,
   Tooltip,
 } from "recharts";
-import { useColors } from "../hooks/widget-hooks";
+import { useColors } from "../hooks";
 import UtilsService from "../services/UtilsService";
 
 type Props = {
