+++ frontend/src/hooks/index.tsx
@@ -5,7 +5,7 @@ import {
   useDashboardVersions,
   useDashboardHistory,
 } from "./dashboard-hooks";
-import { useWidget } from "./widget-hooks";
+import { useWidget, useColors } from "./widget-hooks";
 import { useTopicAreas, useTopicArea } from "./topicarea-hooks";
 import { useHomepage, usePublicHomepage } from "./homepage-hooks";
 import { useSettings, usePublicSettings } from "./settings-hooks";
@@ -30,6 +30,7 @@ export {
   usePublicDashboard,
   useDashboards,
   useWidget,
+  useColors,
   useTopicAreas,
   useTopicArea,
   usePublicHomepage,
